Important bug fix in RNG!
The UNI() macro was pointing incorrectly at the uniformGCC function
which was returning a repeating value.

Also changed here is commenting out of the jsr variable, which is not
used in this code. 2nd arg of zigset has been left in place and has no
effect - it's a dummy arg now. Left in place so that client code will
still compile.