Adds a test that the port container can be indexed
The @port container may be smaller than the number of connections
(num_BRAHMS in this bit of code). This can occur if the user has
imported a csv file for the connection weights with too few
elements. Previously, the code could run off the end of the @port
container and insert junk values when the addImpulse function was
called but without a crash of the program.